requirements.txt
# CSMAR Data Download - Python Dependencies
# Install with: pip install -r requirements.txt

# HTTP client library
urllib3>=2.0.0

# WebSocket support for CSMAR API
websocket>=0.2.1
websocket_client>=1.5.0

# Data manipulation
pandas>=2.0.0

# Pretty table printing for CSMAR results
prettytable>=3.0.0

# Environment variable management
python-dotenv>=1.0.0

# Note: csmarapi must be installed separately from CSMAR website
# Download: https://data.csmar.com/static/python_3.6.0.rar
